[client] vulkan: take color space into account when blending cursor
Currently, the sRGB cursor image is rendered directly into the framebuffer,
resulting in the cursor not matching the host white level when using HDR.

The fixed-function blending capabilities are not really adequate for our
purposes when using HDR. For standard color cursors it could work if we
converted the entire framebuffer from non-linear RGBA10 to linear RGBA16,
but that still doesn't help us with the other cursor types which require
the ability to invert a pixel with reference to the current SDR white
level.

To work around this, add a subpass self-dependency and attach the swapchain
image to the renderpass as both a color attachment and an input attachment.
This allows us to first render the desktop into the color attachment, add a
pipeline barrier to make the changes visible, then render the cursor, where
we can read the rendered desktop image from the input attachment to
manually blend it with the cursor.